pyproject.toml
[project]
name = "gdkvm-20251216"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.7",
    "monai>=1.5.1",
    "numba>=0.62.1",
    "opencv-python>=4.12.0.88",
    "pydantic<2.12",
    "scikit-image>=0.26.0",
    "scipy>=1.16.3",
    "tensorboard>=2.20.0",
    "torch==2.4.1",
    "torchvision==0.19.1",
    "tqdm>=4.67.1",
    "wandb<0.23.1",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" }
]
torchvision = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" }
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

[tool.deptry.package_module_name_map]
"hydra-core" = "hydra"
"opencv-python" = "cv2"
"tensorboard" = "torch.utils.tensorboard"

[tool.deptry.per_rule_ignores]
DEP002 = ["numba", "tensorboard", "pydantic"]
